+++ logc/format.c
@@ -69,9 +69,8 @@ void free_format(struct format *f) {
 
 void log_syslog_format(log_t log, const char *format) {
 	log_allocate(log);
-	if (log->_log->syslog_format)
-		free_format(log->_log->syslog_format);
-	log->_log->syslog_format = parse_format(format);
+	free_format(log->_log->syslog_format);
+	log->_log->syslog_format = format ? parse_format(format) : NULL;
 }
 
 const struct format *default_format() {